[BE] Move linked entity creation to model
